Move tests from gaussian_quadrature.rs into their respective modules
